# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "operationId": "ProductController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductController_findAll_v1",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List products",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{productId}": {
      "get": {
        "operationId": "ProductController_findById_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a product",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductController_update_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid."
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductController_delete_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a product",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Faberon API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Resource was not found."
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateProductRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "manufacturer": {
            "type": "string",
            "example": "HP",
            "minLength": 1,
            "maxLength": 120
          },
          "model": {
            "type": "string",
            "example": "LaserJet M607",
            "minLength": 1,
            "maxLength": 120
          },
          "serialNumber": {
            "type": "string",
            "example": "SN-HP-88213",
            "minLength": 1,
            "maxLength": 120
          },
          "ownership": {
            "type": "string",
            "enum": [
              "by_client",
              "rented"
            ],
            "example": "by_client"
          },
          "address": {
            "type": "string",
            "example": "88 Beachline Dr",
            "maxLength": 240,
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "example": "Front desk",
            "maxLength": 120,
            "nullable": true
          },
          "warrantyDate": {
            "type": "string",
            "format": "date",
            "example": "2027-01-15",
            "nullable": true
          }
        },
        "required": [
          "customerId",
          "typeId",
          "manufacturer",
          "model",
          "serialNumber",
          "ownership"
        ]
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "manufacturer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "warrantyDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "ownership": {
            "type": "string",
            "enum": [
              "by_client",
              "rented"
            ]
          }
        },
        "required": [
          "id",
          "companyId",
          "customerId",
          "typeId",
          "manufacturer",
          "model",
          "serialNumber",
          "address",
          "contactName",
          "warrantyDate",
          "ownership"
        ]
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "manufacturer": {
            "type": "string",
            "example": "HP",
            "minLength": 1,
            "maxLength": 120
          },
          "model": {
            "type": "string",
            "example": "LaserJet M607",
            "minLength": 1,
            "maxLength": 120
          },
          "serialNumber": {
            "type": "string",
            "example": "SN-HP-88213",
            "minLength": 1,
            "maxLength": 120
          },
          "ownership": {
            "type": "string",
            "enum": [
              "by_client",
              "rented"
            ],
            "example": "by_client"
          },
          "address": {
            "type": "string",
            "example": "88 Beachline Dr",
            "maxLength": 240,
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "example": "Front desk",
            "maxLength": 120,
            "nullable": true
          },
          "warrantyDate": {
            "type": "string",
            "format": "date",
            "example": "2027-01-15",
            "nullable": true
          }
        }
      },
...
    }
  }
}